Use :test `QueueAdapter` in specs
The test suite used :inline as the ActiveJob adapter, which performs jobs
inline. This is expensive and unnecessary in most cases.

The `TestAdapter` supports an in-memory queue which makes it easy to test which
jobs have been equeued without processing them, or to process jobs selectively. See:
http://api.rubyonrails.org/classes/ActiveJob/QueueAdapters/TestAdapter.html

We also introduce RSpec example metadata for `:perform_enqueued`, making it easy
to turn on job processing for particular example groups.

Closes #2888