get admin_set_id from existing work if not passed in and work exists
You can end up in this actor from several places.
* new_work_form which goes through #create method and correctly requires admin_set_id attribute
* edit_work_form which will always include an admin_set_id
* manage embargoes -> update which does not include an admin_set_id in which case the current admin set id should not be changed
* manage lease -> update which does not include an admin_set_id in which case the current admin set id should not be changed
* and perhaps others

The proposal in this PR is that if this is an update, the the admin set id should be populated from the work in the env parameter.

NOTE: For validation of embargoes and leases to apply admin set restrictions, the setting of the admin set id on env.attributes must
happen before the actor that validates the restrictions.