Don't crash when a work has no depositor
A work with no
depositor (e.g. https://nurax.curationexperts.com/concern/generic_works/5425k968s)
in its solr document would previously fail on `NoMethodError`.

The tests for the search builders are heavily stubbed, and are in need of a
significant refactor. Since the existing test avoid calling this method, I'm not
adding a test for it here.