Deduplicate exported graphs for nested works
Nested works with hash URIs were previously assumed to be list elements in the
graph exporter for the work presenter. This class is meant to faithfully
replace uris and expand scope of a Fedora resource, but for hash URI's it ended
up refectching and reincluding the base graph using a different replacement
rule.

In the patched version we only follow links when the resource id has a `/`. We
still assume these uris are list elements (which may be wrong). We also fix a
bug where hash URIs got URL encoded, turning hashes into `"%23"`.

Fixes #3103