[refactor] Invert dependencies for `AdminStatsPresenter`
This is a straight refactor to invert the previously hard coded dependencies in
`AdminStatsPresenter`. This allows callers to inject stats services with custom
behavior.

This is particularly motivated by the eager failure behavior of
`Hyrax::Statistics::Depositors::Summary`, which raises an error in the case
that a work has been saved with a `depositor` that can't be found (e.g. the
depositor's user_key has changed, or a depositor string has been saved that does
not have a matching record in the application). Since we don't enforce integrity
of reference for depostor strings, an option for a more fault tolerant stats
service seems desirable. This opens the presenter to that change.

We disable the method parameter size rubocop rule for the `#initialize` method,
since having many optional parameters in an intializer seems like a step forward
from having many hard coded dependencies. Further changes to reduce the
parameter list may be worth exploring.