Handle `nil` responses from `CollectionPresenter#description`
`CollectionPresenter` (and other presenter classes) can return `nil` when
delegating to `#solr_document`. This happens particularly when either the model
does not have a matching property or the indexer fails to index it.

Since `Collection` is not guaranteed to have a `:description` (it's not in
`CoreMetadata`), we need to handle the `nil` case whereever we rely on
`CollectionPresenter#description` in our views. Particularly: apps generated by
Hyrax 1.x do not index `:description` by default, and always return `nil` from
this method.