Assign file_set permissions from work
https://github.com/samvera/hyrax/issues/1421

Remove depositor's edit access to file sets, as previously
assigned via hydra-access-controls. This allows the workflow to control
whether the depositor gets access in the work, and assigns permissions
to the file_sets accordingly.

Modify my/works_search_builder to include works of all states on the
dashboard views for works.