Upgrade browse-everything to latest version
Using the browse-everything gem version 0.16.0
makes Hyrax undeployable (and, presumably,
also breaks the browse-everything functionality)
for those who have a browse-everything config
file that contains symbols. Since Hyrax
has always previous shipped with such a
config file, this would presumably affect
many users of the software.

The fix is to upgrade to 0.16.1, which again allows
symbols in the config file.