Fix homepage controller random spec failure
Fixes https://github.com/samvera/hyrax/issues/3206
This spec was failing randomly, apparently because the search results
are looking for solr field 'date_uploaded_dtsi'. This was not being
loaded, so the value returned here was nil. However the repo was also
not being cleaned before running the spec, so other data could result
in a comparison of nil to an actual value.

It might be ideal to have more documents added for this spec, or to
insure that all works added for specs include uploaded_date, but this
is the first step toward a useful spec (i.e. it actually can find the
GenericWork that is being created for the spec!)