Allow support for `active-fedora` 12
Expand support for `ActiveFedora` to allow use of 12.x releases. This brings us
up to date on current releases.

We explictly add a `Solrizer` dependency, which has been removed from
`ActiveFedora` and inlined to `ActiveFedora::Indexing`. Adding this explictly
retains existing behavior for downstream users. We should remove internal
references to `Solrizer` soon, and remove the dependency altogether on the next
major release.