Extract `VisibilityIntention` class from `InterpretVisibilityActor`
This logic is reasonably complex and reusable, but was previously hidden away in
a nested namespace. The reusable parts of the logic are extracted to a class
with a slightly simplified interface. The existing `Intention` class now
inheirts `VisibilityIntention` and adds parameter sanitization (which doesn't
quite seem to fit in the extracted class).

This is a straight refactor with added unit tests for the new class. The
intention is to reuse this behavior in `hyrax/transactions`.