Remove `#save` of embargoes and leases during InterpretVisibilityActor
This save isn't needed, since the embargo and lease will be saved when the work
is. Having them here is inefficient, and runs the risk of introducing orphaned
objects to the repository.

https://github.com/samvera/hydra-head/issues/226 documents an issue that these
saves are supposed to respond to. Per the tests here, this doesn't seem to be a
problem. I can't see any indication this was ever investigated beyond the
initial report.