Require hyrax/search_state
When attempting to set the admin_set_create_service for the AdminSetsController in the hyrax initializer, you get an uninitialized constant error for Hyrax::SearchState.  It is required in an initializer within the Hyrax engine but for some reason hasn't been loaded by that point in the hyrax initializer.  This commit makes the dependency explicit by requiring the load of search state here.