Add `DestroyWork` transaction, and make it destroy the a work
This is a first step to replacing the `#destroy` stack in the Actor Stack.

A `DestroyWork` transaction is added. It's sole current operation is to destroy
the work if it is persisted. When the work is not persisted, we return a
`Success`; errors are caught and wrapped in a `Failure`.