Recover rolled back minter state on actor stack failure
When the actor stack fails, a database transaction rolls back. When the DB-based
NOID minter is in use, this rolls back the minter state. However, it does *not*
roll back any objects actually created by ActiveFedora in the process of running
the stack. This leads lagged minter state, and a perpetual `Ldp::Conflict`
blocking creation of further objects.

Rather than keeping this lagged minter state, we can catch any errors raised
through the transaction, and update the minter state. To do this, we mint ids
and send HEAD requests to Fedora until we can find an id that does not already
have a Fedora object. This first valid id is wasted, but future mints are
unblocked.

This process only runs in cases where an error has been raised by the
transactions. This is mostly an effective time to remediate, except two caveats:

  - Explict raises of `ActiveRecord::Rollback` are eaten by the transaction, so
    won't recover. Actors, therefore, should not explictly rollback; this is
    probably an anti-pattern to begin with.
  - When an error is raised, but no ActiveFedora objects have been created, we
    are over-eager and end up running the recovery code and wasting the test
    id.

Related to #3128.