Shot in the dark for build failures
In conversation with Chris, RSpec sometimes reports 0 errors, but
the suite returns a non-zero shell response. The CI tool, then
interprets the non-zero response as a failure to build.

My thinking is that the after blocks of rspec may be returning false
for the last run in the suite, and that is interpretted as a non-zero.
This would be sad panda if that were the case, but hey, maybe.

This is very much a stab in the dark.