Refactor `Wings::Valkyrie::ResourceFactory` to use a converter class
Introduces `Wings::ActiveFedoraConverter` and uses it in the
`ResourceFactory`. The converter knows how to get an id and attributes to build
an AF model for a given `ValkyrieResource`.

It currently counts on `#internal_resource` being present, and assumes that the
id is the first entry in `#alternate_ids`. We may need to revisit either of
these assumptions later.