Configure Circle CI to run as a "workflow"
Circle CI supports "workflow" pipelines. This is an initial configuration that
breaks our prior Circle setup down into a workflow to avoid duplication of
work.

The jobs included in the workflow are as follows:

  - `bundle`: this checks out code and installs the `hyrax` level dependencies.
  The output of this step is cached to avoid duplicating work across workflow
  runs, and persists to the workflow's workspace for use by other jobs within
  this workflow run.
  - `build`: generates the Engine Cart application for use by the test suite.
  This depends on the `bundle` job. The output also persisted into the
  workspace.
  - `lint`: runs `rubocop`. This is broken into a separate job so it can run in
  parallel with `build`.
  - `test`: runs `rspec` using the output of `bundle` and `build`. This job is
  configured to run in 4x parallelism, and only after `build` and `lint` have
  succeeded.

A next step might be to fan the workflow out further to include build/test
processes covering supported versions of Ruby and Rails.