[wings] Make Valkyrie a production dependency
We'll start introducing Valkyrie calls into production code shortly, so we need
to make it a production dependency.

Unfortunately, our dependency needs to be 2.0.0+, since the 1.x series has a
hard dependency on `pg`. Passing that dependency along to our users isn't
acceptable; it would make Postgresql a system dependency for every Hyrax
installation. For this reason, we pin to 2.0.0-rc3+.

Since we are ourselves in a beta cycle, this should be okay as long as an actual
2.0.0 release is forthcoming on a short timeline.