Remove Travis CI
Delete the travis configuration and README.md badge.

We got some feedback that `.travis.yml` provides some implicit documentation
about how to run the test suite, which is less easy to understand when reading
the Circle configuration. As a replacement for this, we add a link to the Hyrax
Development Guide's section on running the tests.