[wings] Refactor GenericWorkActorSpec failure example
This test dependend on some deep internals of the BaseActor implementation.

The refactor is more robust. It still depends on internals (the private `#save`
method), but this dependency is isolated to the tested class itself, instead of
reaching into the persistence library.

The goal of this refactor is to remove this test's dependency on ActiveFedora so
we can remove that library from the BaseActor later.