Add method_missing to catch any unconverted methods
NOTE: This will allow methods to function by calling the AF version of the method on the converted resource.  It does not handle the case where you want to pass in Valkyrie resource/id and receive back Valkyrie resource/id.

The function of this is a catch all during the conversion process.