[wings] persist embargo/lease data when embargo/lease_id isn't set
When an embargo or lease is saved in an `ActiveFedora::Base`'s `#embargo` or
`#lease` field, but the `#embargo_id` or `#lease_id` isn't set, we should still
retain the embargo/lease when converting. This supports the pattern for
embargo/lease management that's actually in use for existing AF code. Hopefully,
supporting it will make writing transition code easier.

Visibility is dropped from the `#additional_attributes` to ease integration with
other open branches. The visibility set here is a no-op, already, so whatever
order we choose to remove it should be fine.