[wings] Refactor `id_field` for Hyrax configuratability
Isolate the dependency on `ActiveFedora#id_field` to a configuration. The
implementation in ActiveFedora uses `SOLR_DOCUMENT_ID` if present, and
`ActiveFedora.index_field_mapper.id_field` otherwise.

A next step could be to remove the `ActiveFedora.id_field` call in favor of a
reimplementation of that behavior, referencing `Hyrax.config.index_field_mapper`
instead of ActiveFedora's version.

Related to #3798.