[wings] Add Hyrax finder interface
This interface replaces `ActiveFedora::Base.find` in a way that can support a
clean changeover to Valkyrie.

Existing calls to `ActiveFedora::Base.find` can be swapped for
`Hyrax::ActiveFedoraFinder.find` calls. This class returns the same types, and
raises the same errors as the `AF::Base` version. That can later be swapped for
`Hyrax::ResourceFinder`, which is a light wrapper around
`Valkyrie` `#find_by`.

Some consideration was given to using a Valkyrie custom query, but since the
end-goal for this particular swap is to bundle a globally configured
query adapter and its existing `find_by` method, there seemed to be little
point.

Related to #3790.