[wings] Introduce a converter actors for reworking the actor stack
These are support actors, designed to create a boundary in the stack between
actors that handle valkyrie resources, and code that uses ActiveFedora.

By using both actors adjacent to each other anywhere in the actor stack, a two
way conversion can be achieved. Actors below the boundary will accept Valkyrie
resources as their `env.curation_concern` input and can be designed to act only
on Valkyrie objects. Actors above the boundary can continue to work with
ActiveFedora models.

Alternatively, these actors may be used independantly, to swap input/output for
a specific actor.

The idea is that this boundary can begin at the bottom of the stack, and actors
can be reworked or replaced to move the boundary up the stack.