[wings] Setup Valkyrie Solr index
Adds configuration for the Valkyrie Solr index. The configuration is added to
via `config/valkyrie_index.yml`, which has defaults for

This introduces new environment variables:

  - `VALKYRIE_SOLR_HOST`
  - `VALKYRIE_SOLR_PORT`
  - `VALKYRIE_SOLR_CORE`

In keeping with the behavior of other environment-based configuration, we
provide defaults--even in production--so the introduction of these variables
doesn't require immediate action.

We use the blacklight configured host and port if no configuration is present.