Manually control regeneration of the enginecart application in CI
The circle cache key had previously been trying to capture each file that might
change, causing a need to regenerate the test application. However, its strategy
was incomplete.

In general, we need to regenerate the test app whenever anything in
`lib/generators` changes. We should also regenerate whenever an initializer
changes (to test the initializer's interaction with the generators). There are
probably other cases where we should force a regen.

Rather than trying checksum all the possible file changes, we checksum a new
`.regen` file. For now, this file just contains an incrementing counter, but we
could consider later adding a script which writes to the file deterministically
in a way that will change whenever we want to regen the app.

This isn't perfect, but it seems like a step in the right direction.