[wings] Refactor `AttachMembersActor` for Valkyrie membership adding
Add members using Valkyrie.

First, we check the the Ability's permissions against the passed in `id` instead
of a specific object. Then we grab the `Valkyrie::Resource` for the env`s
`#curation_concern`, add the new member, and convert back to `ActiveFedora` to
reset the `#curation_concern`.

This removes the need for an explict `#find` in the private `#add` method, but
this find is, in fact, retained in the `Wings::ActiveFedora::Converter`. Further
refactors may be possible to do this addition directly

All the changes here are isolated to a private method, so the Valkyrie resources
won't leak into public behavior.