[wings] Use Valkyrie for destroys and inserts in AttachMembersActor
This is a major "valkyrizing" refactor for `AttachMembersActor`. It inlines the
old private `#add` and `#remove` methods to
`#assign_nested_attributes_for_collection` (also private). This method now
casts `env.curation_concern` to a `Valkyrie::Resource` once, edits the
membership in that form, then casts back to `ActiveFedora`. This allows the
to the member ids lists to happen in two successive changes (inserts, then
deletes), rather than processing the attributes successively.

As a next step, we can support `Valkyrie::Resource` directly in this actor by
skipping both cast actions if `env.curation_concern` is already a
`Valkyrie::Resource`.