Set destroyed flag on `FileSetActor#file_set` after destroying
After using a `Valkyrie` delete, we need to adjust the state of the in-memory
`file_set` to avoid an unneeded round-trip to Fedora.

Additionally, we introduce a small monkey-patch to `ActiveFedora::Core#reload`
to ensure that, even if we do send the unneeded request, we raise the correct
error and conform to `ActiveFedora`'s desired state.

We'll submit a PR to `active_fedora` to replicate this fail-safe behavior in the
mainline.