Refactor `VisibilityJob` to be single responsibility
Move the resposibility for actually changing visibility out of the Job, allowing
job code to focus on asynchronous processing. An abstract `VisibilityPropogator`
with a factory method (`for(source: my_work)`) is introduced, and the visibility
copy code is moved into a concrete implementation.

This prepares for the introduction of an additional `VisibilityPropogator` that
supports `Hyrax::Resource` objects.