[wings] Introduce valkyrie-native Embargos
Adds a `Hyrax::Embargo` as the valkyrie base class for embargos. The existing
Wings embargo handling is updated to use this base class.

A `NullEmbargo` is introduced as a static empty/inactive embargo. Whenever an
embargo is absent in Valkyrie driven code, we should use this `NullEmbargo`
class to represent a non-existent embargo on the object.

Additionally, an `EmbargoManager` is added to handle operations on the Embargo
as it relates to a given `Hyrax::Resource`.

Connected to #3869.