Don't short-circut actor Environment when interpreting visibility
`InterpretVisibilityActor` previously short-circuited `env` by initializing a
completely new copy for use further down the stack. Changes to the environment
after that point would fail to have any impact further up the stack, where
actors are already using the older `env`.

The original reason the `InterpretVisibilityActor` did this was to alter the
attributes in batch; i.e. this was a work around for `#attributes=`. Now that we
support that method, we can use it instead.

The `CreateFilesActor`, by contrast, was counting on those attributes not to
change. Since any actor further down the stack may change them, it's better to
cache them immediately.