[wings] Use the model's id minting logic in the transformer
The ActiveFedora model provides an `#assign_id` method, which can be used
directly instead of reaching into the id minter for assigning ids. This
reinstates compatibility between wings and models that are using Hyrax's
prescribed interfaces for minting identifiers.

By default, this only mints ids when the noid minter is enabled.