Add a custom query for finding AccessControl objects
Given a resource, we want to be able to find an AccessControl object for it.

For a typical valkyrie adapter, this is an inverse reference lookup on the
`AccessControl#access_to` property. An initial implementation does a broad
reverse lookup on `#access_to`, then filters the result classes. It assumes the
first `AccessControl` found is the correct one.

For Wings, the story is a bit different. `Hydra::AccessControl` resources don't
know what their target is. When rebuilding valkyrie resources from
`Hydra::AccessControl`/AF objects, we inspect the permissions to assign
`#access_to`, but we can't use this for lookup. Instead, we can rely on the
`ActiveFedora`/`Hydra::AccessControls` pattern of looking up the
`Hydra::AccessControl` object from its target (via its `belongs_to` reference).