[wings] Remove `convert_class_name_to_valkyrie_resource_class`
This was an early addition to Wings, and predates much of the query work. at
this point, objects round trip just fine without the special class casting
handling. For certain adapters (e.g. the memory adapter) the call to
`internal_resource` results in a `NoMethodError` on `Object`, so this seems like
a good time to remove the superflouous code.