Delegate read/edit groups and visibility to `PermissionManager`
Allow a `Hyrax::Resource` to cache a `PermissionManager` (and therefore an ACL
`ChangeSet`). Delegate current `#read_groups=` and similar management to it, and
also use it in `VisibilityReader` and `VisibilityWriter`. This allows each
Resource to maintain its own current permission state across the various access
lenses (visibility, `*_users/*groups`, and permissions) in a single place.

When we convert from an `ActiveFedora` object, we ensure that the current
permissions are copied down to the Resource. Notably we don't do this in the
other direction; our implementation of permissions in Valkyrie treats them as
separate resources, pointing back to repository Objects. Changing an ACL, and
then saving the object it points to has no effect: you need to save the ACL
explictly.