Upgrade Valkyrie to 2.0
The main impact of this is that `Dry::Validations` and `Dry::Struct` are
updated. Neither is backward compatible.

The `WorkflowSchema` is rewritten using the new syntax, and it is now slightly
more strict. It expects `notifications:` to have names that can be constantized
to a `Class`, not just to any registered constant.

`Wings::ConverterValueMapper` has to handle ids and nested structs/`Valkyrie::Resources`
slightly more aggressively, since types are more agressively coerced, and aren't
squashed into hashes when nested in a parent's `#attributes`.