prefer pattern of setting use_valkyrie in the initializer
This pattern cannot generally be used for class methods.  For instance, the query service has several class methods that accept `use_valkyrie:` as a parameter to determine if the result should return a valkyrie resource or an AF object.  Which is preferred is on a case by case usage of the caller.

The exception to the class methods pattern is solr_service where class methods are delegated to instance methods.  Whether or not to use_valkyrie is determined by `Hyrax.config.query_index_from_valkyrie`.