create methods for attaching to work using valkyrie and using active fedora
There are multiple steps in the attach_to_work method that must change to work with valkyrie.  If each step is separated into a method, the conversion between active fedora and valkyrie would have to happen multiple times.  So in this case, the entire method is written once for active fedora and again for valkyrie.  A switching ‘perform’ method is used to determine which to call.

For new works to be able to complete the attach_to_work process when using Valkyrie, the updated work needs to be returned to the caller.  This is because the passed in work does not have an id and the resource save process returns a new instance of the work with the id.  The original work object has no way to reload the values saved by the valkyrie persister.

Marked pending the test for versions as the valkyrie resource is not maintaining the version.