failing test for reload new af object failure
### AF Testing Pattern in use in Hyrax:
* build an af_object and don’t save
* call a method that modifies the af_object and saves
* in test, reload af_object and test changes

### Challenge when using valkyrie

* build an af_object and don’t save
* call a method that creates a valkyrie resource from the af_object, modifies the resource and saves, converts the resource back to an af_object (this has a different object_id from the one passed in)
* in test, reload original af_object passed to the method and test changes

FAILS because the original af_object does not have the same `object_id` as the af_object converted from the saved resource.  And because it wasn’t saved, it does not have the same `af_object.id` value.  The one in the test has `af_object.id=nil`.

So the question is how to address this as we continue converting code and tests to be valkyrized.