Provide a Valkyrie changeset factory
For any given Valkyrie model, we want a factory for an appropriate `ChangeSet`,
with attribute accessors and a `#sync` method.

We provide this with a `Hyrax::ChangeSet.for(my_resource)` factory. This builds
the changeset class on the fly, applying the fields from the resource
schema. Then a new instance of the class is initialized to handle the `ChangeSet`
for the specific object.

As future work, we could consider caching the changeset classes. For now, it
seems best to watch and wait--avoiding premature optimization.