Introduce a `SimpleWork` for testing, and a factory for it
This "Simple Work" is meant for testing native-Valkyrie works. It has only the
core Hyrax metadata for now, but should be extended with the minumum Work
model (members, etc...; i.e. it should inherit `Hyrax::Work` or something along
those lines).

A factory is added with some basic visibility traits.

For Wings support, `SimpleWorkLegacy`, a work with `CoreMetadata` and
`WorkBehavior` is provided.