Qualify references to the application's `Collection` class
Use `::Collection` when we mean to refer to the application-layer `Collection`
class. This ensures correct reference even if a class named `Collection` is
defined in another namespace.

We already provide this kind of disambiguation for `FileSet`. In general, we
should use `::`-prefixed names for all application-level names we want to
reference from the engine (inverted dependencies on the application). Failing to
do this means any new name defined in an in-scope namespace will resolve
preferentially, over the desired application class/module. If we intend to refer
to an application constant, we should do so with care.

For now, this will keep us from defining `Hyrax::Collection`, since the fallout
for adopters is a bit unpredictable. Still, it's good to clean up after our
mess, here.