[wings] Avoid special handling for values in `ConverterValueMapper`
Refactor `Wings::ConverterValueMapper` to use the `::Valkyrie::ValueMapper`
chain in place of all prior special handling. Where we used to delete attributes
manually, now we capture and return `nil` for them during conversion.

This helps clarify the classes of special behavior that we have, and opens the
door to refined handling for edge cases.
Remove the `ActiveFedoraAttributes#filter_attributes` method.