Start to generalize Work Controllers for Valkyrie
Working under the hypothesis that `WorkControllerBehavior` can reasonably
support Valkyrie native Works as well as ActiveFedora ones, begin to test
behavior for Valkyrie works in `works_controller_behavior_spec.rb`.

In the past, we've tested controller behavior using generated work
controllers (`GenericWorkController`) in the test application. Locating the
tests for this sharable module in a file named for a specific implementation
makes it hard to find, and encourages sloppy testing. So we are shifting away
from this approach, and instead bootstrapping a controller and routes directly
for use in testing---if we need additional controllers to test other code paths,
they should be easy to add following this pattern.

A `ValkyrieCanCanAdapter` is added to resolve Resource objects from
ids. Inheriting `CanCan::ModelAdapters::AbstractAdapter` automatically registers
this adapter for use based on `.for_class?`.

A simple version of `Hyrax::Ability` support is added via a specialized
`#extract_subjects` case: whenever the `subject` is a `Hyrax::Resource`, cast it
to an ID and check permissions based on this. This will often (always? )involve
querying permissions from Solr within `Ability`, but should work for most
ability checks. As usage in Controllers grows, we'll need to support Valkyrie
native ability checking.