Implement Valkyrie native Collections as `PcdmCollection`
Add a PCDM style collection in native Valkyrie, with support for
membership. Extract Work membership tests into shared examples for use across
PCDM/Works models.

We name the Collection model `Hyrax::PcdmCollection` to avoid application layer
collisions with `::Collection` when referring to it from a Hyrax module. That
is: we don't want to introduce a `Hyrax::Collection` without warning adopters.