Add a Valkyrie-based Admin Set model (`AdministrativeSet`)
Implement a basic Valkryie Admin Set model. Add support for admin sets to
Valkyrie native Work models.

We played with the idea of making `AdministrativeSet` a subclass of
`PcdmCollection` but the assumption of ordered membership in collection is
problematic, here. We don't want order in Admin Sets, and the AF models won't
support it.