Deprecate `PowerConverter` based `to_sipity_action`
`PowerConverter` is a community-maintained gem, but is not in the Samvera
namespace. Depending on it under these circumstances is not in keeping with our
policy.

Rather than simply inline the `PowerConverter` class, this proposes pushing in
the direction a more generic ruby pattern that inspired
it. http://devblog.avdi.org/2012/05/07/a-ruby-conversion-idiom/