Compile javascript into an NPM package
This will allow us to support webpacker as well as continue support of sprockets in the manner used by Blacklight.

The javascript is built by npm from sources in `app/javascript` into a bundle in `app/assets/javascripts/hyrax.js`. This file should not be edited by hand as any changes would be overwritten. When any of the javascript components in the gem are changed, this bundle should be rebuild with the following steps:

1. [Install npm](https://www.npmjs.com/get-npm)
1. run `npm install` to download dependencies
1. run `npm run js-compile-bundle` to build the bundle
1. run `npm publish` to push the javascript package to https://npmjs.org/package/hyrax-frontend