Use the `Hyrax::TimeService` for events
`Hyrax::Event` usually wants to be created "now". For Hyrax, "now" generally
means `Hyrax::TimeService.time_in_utc`.

We switch to using this method, and add a utility method for this. The method is
convenient for users, and extracts behavior from `EventJob` where it's not as
visibile to users who might be trying to understand how to use the Event.