Use a factory method (`.for`) for to get valkyrie indexers
We'll want to get different indexers for different kinds of resources, so we can
prepare them for indexing according to a customized schema. Switching to a
factory method instead of an initializer allows us to create a new indexer
instance based on the properties of the resource.

In Hyrax, we'll likely build from the factory depending on class, but
applications can easily override to build depending on schema/attributes,
present values, etc..., by providing a different `ValkyrieIndexer` to their
`IndexingAdapter`, with a custem `.for` implementation.

Connected to #3800.