Test publisher/listeners for `GenericWorkActor`, instead of callbacks
The previous tests were tightly coupled to the specific arguments of the
callbacks, which will change/expand as part of deprecation. Rather than continue
to test the old behavior so strictly, we transition to testing with a listener.

The spy listener is less tightly coupled to the publishing mechanism, and
instead tests the expected behavior for clients.