Test indexing basic metadata for Hyrax
Add a shared spec for indexing basic metadata using the new
`Hyrax::Indexer`. Since none of the hyrax core models include basic metadata by
default (yet?), the tests are run as an extension on `ValkyrieWorkIndexer`. This
feels natural/appropriate, since it guarantees that the extension for tha class
works as expected.