Move wings setup to its own file
Some of this setup is a bit fragile when called during app generation. By
isolating it, we can handle errors more carefully. We always reraise when in
production mode, since that means we aren't generating the app---we want to fail
eagerly in this case to avoid rampant runtime errors.

This replaces a more narrow error handling approach in `ModelRegistry`.