Write ACL changes to valkyrie index with a listener
When an ACL change event is published, write to the currently configured
Hyarx (valkyrie) indexing adapter. This ensures that when ACLs are changed
independently of the object, the permissions cached in Solr will stay up to
date.

Hyrax relies on permissions to be cached correctly in the index for various
`Ability` behaviors. This is perhaps not ideal for a long-term valkyrie future;
`ActiveFedora` basically requires this of us, though, unless we're willing to
take the performance hit of looking up the ACLs from Fedora per-request.

Note that this is presently a no-op by default, since we use a null indexing
adapter.