Add tests for `WorksControllerBehavior#show`, `#create`, & `#new`
Adds module level tests for `WorksControllerBehavior#show` using a
controller registered for a Valkyrie model. A minor changes to `GraphExporter`
is needed to make some of these tests pass (the changed branch is otherwise
untested).

The outline of tests for `#create` and `#new` are also put in place, but these
require more work (on Actors and Forms, respectively) to pass. Accordingly, they
are left as pending tests.