Add a `ChildTypes` to manage selection display during work attachment
With `ActiveFedora` models, we use `MyWork.valid_child_concerns` to determine
which types the user can select to attach to an existing work.

Since this is a display-only issue, we break this out into a new service. For
legacy models, we delegate the configuration back to the model, avoiding any
need for configuration changes.

For Valkyrie models, we provide a default config (a work can nest works of its
own type), but punt on configuration issues. There are other configuration
matters to discuss.