Support JSON serialization for Valkyrie in WorksControllerBehavior
Add support for JSON rendering in `WorksControllerBehavior#show` when a valkyrie
concern type is specified. We do this by casting to the legacy model in the
view. This ensures we can keep exactly the same JSON output as we transition.

We'll need to rework the JSON APIs (such as they are) in order to turn off
Wings.