Add a `hyrax:work_resource` generator for valkyrie native works
The new generator creates a valkyrie-driven `Hyrax::Work` subclass, with a
controller and a spec file. The specs include the Hyrax shared examples (and the
valkyrie shared examples) by default.

Since a custom search builder is needed to ensure Valkyrie works can be found by
their legacy `ActiveFedora` model names, we include a class builder for that,
and register the custom class with the generated `Controller`. This can be
removed when Wings is deleted; applications that are ready to move to
valkyrie-only index behavior will need to drop this search builder in favor of
one that uses the valkyrie index at a later date.